+++ go.mod
@@ -5,7 +5,6 @@ go 1.17
 require (
 	github.com/gin-gonic/gin v1.7.7
 	github.com/joho/godotenv v1.4.0
-	github.com/rs/xid v1.3.0
 	go.mongodb.org/mongo-driver v1.8.3
 )
